extract Clock interface. Make ClockImpl package private
