+++ lib/src/test/java/com/auth0/jwt/ClockImplTest.java
@@ -1,19 +1,19 @@
 package com.auth0.jwt;
 
+import com.auth0.jwt.interfaces.Clock;
 import org.junit.Test;
 
 import java.util.Date;
 
-import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
 import static org.junit.Assert.*;
 
-public class ClockTest {
+public class ClockImplTest {
 
     @Test
     public void shouldGetToday() throws Exception{
-        Clock clock = new Clock();
+        Clock clock = new ClockImpl();
         Date clockToday = clock.getToday();
         assertThat(clockToday, is(notNullValue()));
     }
